Updated error message view and Asp.net timeout value
Added display of error message if token is expired, and requiring user
to re-authenticate.   Improved the appearance of the table view in User
Index view.
Increase session state timeout

The default timeout in ASP.Net is 20 minutes, so the cached access token
and refresh token were being cleared after 20 minutes, leaving no
opportunity to observe the access token expire and the refresh token
used to
get a new access token, therefore we increased the session timeout value
in the webconfig file.