Updating Startup.Auth.cs to support Certificate Credentials
This is an update to the Startup.Auth code to allow support for authentication using Certificate Credential.
Web.Config is updated to include a new <appSetting> called "ida:CertName" which can contain the CN of the certificate the user wants to use to authenticate. If the CertName value is empty, authentication with App Key will occur. If the CertName value is populated (certName.Length != 0), it will attempt to authenticate using a Client Certificate.