Updating Web.config to support Certificate Credentials
This is an update to add a new <appSetting> for "ida:CertName" which allows a user to use a certificate to request an access token.
This corresponds to a change in Startup.Auth.cs which checks if the CertName value is populated, i.e. ( certName.Length != 0). If it is, the application will attempt to authenticate using a client certificate. If the CertName value is empty, it will use the AppKey to authenticate instead.